+++ pre-processing-console/Models/TermFrequence.model.cs
@@ -6,7 +6,7 @@ namespace pre_processing_console.models
     public class TermFrequenceGroupTrainingModel
     {
         public Guid Id { get; set; }
-        public List<KeyValuePair<int, int>> averageWordsCount { get; set; }
+        public List<KeyValuePair<int, double>> averageWordsCount { get; set; }
         public int keywordsLimit { get; set; }
         public double tolerance { get; set; }
