+++ pre-processing-console/Models/preprocessedDocument.model.cs
@@ -8,6 +8,6 @@ namespace pre_processing_console.models
         public string fileName { get; set; }
         public string fileId { get; set; }
         public string fullText { get; set; }
-        public List<KeyValuePair<int, int>> keywordsCount { get; set; }
+        public List<KeyValuePair<int, double>> keywordsCount { get; set; }
     }
 }
